+++ bundle/src/main/java/com/adobe/acs/commons/redirects/filter/RedirectFilter.java
@@ -419,14 +419,15 @@ public class RedirectFilter extends AnnotatedStandardMBean
             if (mapUrls()) {
                 location = mapUrl(location, slingRequest);
             }
-            if(preserveQueryString) {
-                String queryString = slingRequest.getQueryString();
-                if (queryString != null) {
-                    location = preserveQueryString(location, queryString);
-                }
-            }
             if(urlAdjuster != null){
                 location = urlAdjuster.adjust(slingRequest, location);
+            }
+        }
+        if (preserveQueryString) {
+
+            String queryString = slingRequest.getQueryString();
+            if (queryString != null) {
+                location = preserveQueryString(location, queryString);
             }
         }
         return location;