+++ bundle/src/main/java/com/adobe/acs/commons/wcm/vanity/WrappedRequestPathInfoWrapper.java
@@ -2,7 +2,7 @@
  * #%L
  * ACS AEM Commons Bundle
  * %%
- * Copyright (C) 2013 Adobe
+ * Copyright (C) 2017 Adobe
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,8 +17,7 @@
  * limitations under the License.
  * #L%
  */
+package com.adobe.acs.commons.wcm.vanity;
 
-@Version("2.3.0")
-package com.adobe.acs.commons.synth;
-
-import org.osgi.annotation.versioning.Version;
+public interface WrappedRequestPathInfoWrapper {
+}