+++ CHANGELOG.md
@@ -13,11 +13,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 - #2730 - Optimized Shared Component Properties feature with request attribute caching and injector reliance on BVP
 
+### Changed
+
+- #2742 - Provided a way to author localized titles for Generic Lists.
+
 ## 5.0.14 - 2021-10-20
 
 ### Fixed
 
-- #2742 - Provided a way to author localized titles for Generic Lists.
 - #2704 - Fixed issue with MCP report generation throwing an exception, and fixed some minor UI issues on AEM SDK (added BG color)
 - #2716 - Fixed issue with Shared Component Properties Bindings Values Provider facing lock contention
 - #2718 - Fixes CM Code Quality Pipeline failure caused by TestMarketoInterfaces and Jacoco instrumentation