+++ CHANGELOG.md
@@ -9,12 +9,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->
 [unreleased changes details]: https://github.com/Adobe-Consulting-Services/acs-aem-commons/compare/acs-aem-commons-5.0.14...HEAD
 
+## 5.1.0 - 2021-12-13
+
 ### Added
 
 - #2741 - Add a new render condition iscurrentusermemberof - A condition that evaluates to true, if the current user is a member of required groups or an admin.
 
 ### Fixed
-
 
 - #2749 and #2488: Manage Controlled Processes does not show any process fixed via #2751
 - #2337 - Marketo form null on publish fixed via #2758